moving routes under the same module
